Add ID (inline) value class

+++ features/routines/src/main/java/com/enricog/routines/detail/summary/ui_components/RoutineSummaryScene.kt
@@ -62,7 +62,7 @@ internal fun RoutineSummaryScene(
                     when (item) {
                         is RoutineSummaryItem.RoutineInfo -> item.hashCode()
                         is RoutineSummaryItem.SegmentSectionTitle -> item.hashCode()
-                        is RoutineSummaryItem.SegmentItem -> item.segment.id
+                        is RoutineSummaryItem.SegmentItem -> item.segment.id.toLong()
                         RoutineSummaryItem.Space -> item.hashCode()
                     }.exhaustive
                 }
